Add the backspace button to the list of memory buttons
The backspace button seems not to be added to any set of buttons,
reulting in incoherent font color.
This commit adds it to the memory buttons, like the clear button,
making the font color coherent with the rest of the buttons and
background and foreground colors to be set in the proper section.